+++ packages/stream_chat_test/lib/stream_chat_test.dart
@@ -15,6 +15,7 @@ export 'package:test/test.dart';
 
 // Helpers
 export 'src/helpers/api_mocker_mixin.dart';
+export 'src/helpers/matchers.dart';
 export 'src/helpers/mocks.dart';
 export 'src/helpers/test_data.dart';